blockbusted_django_render_polling/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
psycopg[binary]==3.2.9
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.10.0
